% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncourt_query.R
\name{connect_oncourt_db}
\alias{connect_oncourt_db}
\title{Oncourt database connection}
\usage{
connect_oncourt_db(
  db_path = Sys.getenv("oncourt_path"),
  pw = Sys.getenv("oncourt_pwd")
)
}
\arguments{
\item{db_path}{path to Oncourt Access database file}

\item{pw}{password (contact with OnCourt administrators)}
}
\value{
connector to database
}
\description{
Before trying connect to OnCourt database, to simplify the process, please set-up environment variables by using usethis::edit_r_environ()
.Two variables are needed :
\itemize{
\item \emph{oncourt_path} - path to Oncourt database file, e.g. on Windows: C:/Program Files (x86)/OnCourt/OnCourt.mdb
\item \emph{oncourt_pwd}  - password to database
}
}
