% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncourt_query.R
\name{query_oncourt_matches}
\alias{query_oncourt_matches}
\title{Get historical data}
\usage{
query_oncourt_matches(
  type,
  date_min = as.Date(Sys.time()) - 7 * 2,
  date_max = as.Date(Sys.time()),
  con = cn_access
)
}
\arguments{
\item{type}{wta or atp}

\item{date_min}{lower bound of date}

\item{date_max}{upper bound of date, default today's date}

\item{con}{connector to database}
}
\value{
data frame with historical data of particual players
\itemize{
\item \emph{type}     - atp or wta
\item \emph{type2}    - single or double
\item etc...
}
}
\description{
Get historical matches constrained by date
}
