% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncourt_query.R
\name{query_oncourt_playerdata}
\alias{query_oncourt_playerdata}
\title{Get historical data by individual player}
\usage{
query_oncourt_playerdata(
  type,
  date_min = as.Date(Sys.time()) - 365 * 4,
  date_max = as.Date(Sys.time()),
  players_id = c(45854, 54663),
  con = cn_access
)
}
\arguments{
\item{type}{wta or atp}

\item{date_min}{lower bound of date}

\item{date_max}{upper bound of date, default today's date}

\item{players_id}{ID's of players to querying, e.g. column ID_P_R from query_oncourt_playerranking()}

\item{con}{connector to database}
}
\value{
data frame with historical data of particual players
\itemize{
\item \emph{type}     - atp or wta
\item \emph{score}    - win or lost from perspective of player 1 (P1)
\item \emph{ACES_1}   - aces of player 1
\item \emph{DF_1}     - double faults of player 1
\item \emph{TPW_1}    - total points won
\item \emph{TPW}      - TPW_1 + TPW_2
\item \emph{TIME_G}   - length of match, in minutes
\item etc...
}
}
\description{
Get historical matches with statistics
}
