% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncourt_query.R
\name{query_oncourt_playerranking}
\alias{query_oncourt_playerranking}
\title{Get player ranking info}
\usage{
query_oncourt_playerranking(type, from = 730, con = cn_access)
}
\arguments{
\item{type}{wta or atp}

\item{from}{lower bound of ranking release date expressed in integer number, e.g. 730, from 2 years back to now}

\item{con}{connector to database}
}
\value{
data frame with ranking data
\itemize{
\item \emph{DATE_R}  - date release of ranking
\item \emph{ID_P_R}  - ID of player
\item \emph{POINT_R} - points
\item \emph{POS_R}   - position
\item \emph{DATE_P}  - date of birth
\item etc...
}
}
\description{
Get actual and historical ranking informations
}
