% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncourt_query.R
\name{query_oncourt_schedule}
\alias{query_oncourt_schedule}
\title{Get schedule}
\usage{
query_oncourt_schedule(type, con = cn_access)
}
\arguments{
\item{type}{wta or atp}

\item{con}{connector to database}
}
\value{
data frame with schedule data
\itemize{
\item \emph{type}     - atp or wta
\item \emph{type2}    - single or double
\item \emph{NAME_T}   - tournament
\item \emph{NAME_R}   - round
\item \emph{NAME_C}   - surface
\item \emph{COUNTRY_T}  - country
\item etc...
}
}
\description{
Get actual schedule of single & double matches
}
